requirements.txt
django==3.1.6
django-allauth==0.44.0
django-auth-ldap==2.2.0
django-compressor==2.4
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.8
psycopg2-binary==2.8.6
python-dotenv==0.15.0
